finagle-base-http: Configuration methods on Cookie to create a new Cookie with a field set
Summary: Problem

Sometimes when building a Cookie, we have Optional
values for configuration parameters. We are deprecating
the mutability of Cookie, so doing something like
`optDomain.forEach(cookie.domain = _)` will no longer
work. It would be useful to be able to configure an existing
cookie with a new param.

Solution

For each configurable field on `Cookie`, introduce a corresponding
method that creates a new `Cookie` with that field set.

JIRA Issues: CSL-5080

Differential Revision: https://phabricator.twitter.biz/D83226