finagle-mysql: Move cursor method to Client, remove Cursors trait
Summary: Problem

Running multiple SQL statements inside a transaction is an important and
fundamental feature of an RDBMS.  Using the `cursor` method is a valid and
supported way of querying, just as `select` is, but today you can only access
the `cursor` method with a `Client with Cursors` instance.  This means you
cannot use that with the `transaction` method, which requires a `Client`.
This limitation becomes important in use cases where one needs to run two or
more queries with read consistency, and one of the queries is a cursored query.
Without a transaction, the statements are not guaranteed to have read
consistency.  Also, when reading from a read-only database via a load balancer,
each query could be directed to a different host, which may not be in the same
replication state.

Solution

Moved `cursor` into the `Client` trait, and removed the `Cursors` trait.
Removed all references to `Cursors` trait in the source tree.
Added stub and pass-through `cursor` methods to test code implementations of
`Client`.

Result

Users will now be able to use cursors inside transactions.

JIRA Issues: CSL-4828

Differential Revision: https://phabricator.twitter.biz/D91789