DeadlineFilter: Deprivatize the DeadlineFilter class
Summary: Problem/Solution
With access only to the DeadlineFilter object, it can only be pushed on to a
stack.  It is useful to have direct access to the class so that it can be
chained as a regular filter.

Result
A new DeadlineFilter can be created and chained with other Filters and Services

Differential Revision: https://phabricator.twitter.biz/D94517