Document the potential risk in allowing the dtab-local header
Summary: Problem:
The default Http.server stack allows any request to set dtab-local. This
could be a security hole for front-end OR backend services. If a bad
actor set this header (perhaps via a browser plugin), they may
be able to force requests to unsafe locations and retrieve sensitive information.

Solution:
Document the ServerDtabContextFilter and update documentation with
tips on how to disable.

Differential Revision: https://phabricator.twitter.biz/D95380