fingle-netty4: Fix race in Netty4PushChannelHandle
Summary: Problem

When Netty4 sets a channel to inactive, it removes all handlers from the
pipeline. This can race with the `DelayStage.installSessionDriver` method
which tries to remove the DelayStage, but ends up throwing a
NoSuchElementException and subsequently fails to update a Promise,
resulting in a TimeoutException.

Solution

Check if the DelayStage has been removed from the pipeline before
attempting to continue.

JIRA Issues: CSL-5287

Differential Revision: https://phabricator.twitter.biz/D95726