[scala-strict-deps] Remove strict_deps=True for all remaining targets in source
Summary: This is an automated change. We’re not expecting you to review the individual changes to the BUILD files you own, but we want you to be aware that they are happening.

This patch removes strict_deps=True for the remaining targets in Source because it is enabled by default globally.

The script only removes the explicit setting of `strict_deps=True` in BUILD files, and does not change any other aspect of the build file. For more information about strict_deps, please refer to http://go/strict-deps.

Strict deps restricts what is on the compile classpath to improve compile times. This means you spend less time waiting for your builds. However, with this change, you need to remember that transitive dependencies don't work so if you need something to compile code, you need to add it directly to your dependencies.

TBR=true

NO_USER_HOOK=1

Differential Revision: https://phabricator.twitter.biz/D97364