finagle-stats: Underlying metrics maps are shared by default
Summary: Problem

The underlying metrics maps (stats, gauges, counters) are not shared
across Metrics instances. This means that when there are multiple
stats receivers, counters and stats (which are created per-registry)
are created twice.

Solution

Make metrics maps shared by default, and add a function, `createDetached`,
to Metrics, which uses new maps. This restores the behavior prior to
D73497, which removed the ability to created "detached" Metrics and
switched all usages to be undetached.

JIRA Issues: CSL-5343

Differential Revision: https://phabricator.twitter.biz/D99005