finagle-serversets: c.t.f.serverser2.Stabilizer refactor which includes fix for duplicate shardids
Summary: Problem:

Zk2Resolver uses c.t.f.serverser2.Stabilizer utility to dampen address changes for the observers
so that the they are not affected by transient or flapping states from the external system which
supplies the source addresses. It buffers and batches consecutive sequence of bound addresses so
that transient changes to individual addresses are not propagated immediately. It uses two timers
(Epochs) to manage address removals and batching. The removals are only sent out once in
removal Epoch (defaults to 40 seconds for Zk2Resolver). Address additions are published every 10
seconds. However, as a side effect, when backend nodes are started it sends out
duplicate hosts for the same shards and the caller gets stuck with them upto 40 seconds. This
specially breaks clients that need shard uniqueness for partitioned clusteres.

Solution:

Refactored and cleaned up Stabilizer completely. It now does the failure and address stabilization
using the same epoch. Transient issues that result in failed resolution are ignored and previously
known healthy state is used. For shielding the observers from volatile sources which churn the
underlying addresses, it buffers and batches addresses in the same epoch. As a result it sends
out address additions every epoch and holds the removals for one extra epoch so that flappiness
in addresses gets a chance to recover. When shardids are present it dedupes the new addresses
against the bufferred ones, to fix the duplicate shardid issue. We eliminated the old time
computation for batching changes to get rid of needless complexity. Earlier if a previously
resolved address becomes unresolvable that was hidden from the clients. Now it propages
Addr.Pending to the observers. Also upon failures, removes were completely dropped. Now
it keeps promoting the last known healthy state, because those updates came in before we
encountered the failure.

Result:

This fixes the duplicate shard ids issue and makes the code simpler and easier to maintain.

JIRA Issues: FES-3180

Differential Revision: https://phabricator.twitter.biz/D100235