finagle-core: Metrics for assessing consistency of peerset and serversets
Summary: Problem

We don't have a good handle as to whether our load balancers are getting
consistent views of the serverset and we don't know if all clients think
they are in the same peerset.

Solution

Add a gauge that tracks a hash of the serverset and a guage that tracks
the size of the peerset.

JIRA Issues: CSL-5380

Differential Revision: https://phabricator.twitter.biz/D101365