finagle MethodBuilder: Add logging of unsuccessful requests
Summary: Problem

When there are failures on specific methods, it is difficult for
service owners to know what the cause of them are as failure counters
are not kept in order to keep the metrics to a minimum.

Solution

Add a new Filter that logs unsuccessful results at debug level.

Result

Service owners have better tools to investigate the cause of problems.

JIRA Issues: csl-5182

Differential Revision: https://phabricator.twitter.biz/D102152