finagle-netty4: TrustManager.Insecure Works with Native Engines
Summary: Problem

Specifying `TrustManager.Insecure` forces the underlying `SslProvider`
to be set to use the JDK version for the Netty 4 engine factories. At
the time it was written, the provider was forced due to a restriction
within Netty, which no longer exists.

Solution

Remove the forcing of the JDK `SslProvider` when `TrustManager.Insecure`
is specified.

Result

`tlsWithoutValidation` can be used with native engines, and for
protocols like HTTP/2 that require a native engine at this time.

JIRA Issues: CSL-5406

Differential Revision: https://phabricator.twitter.biz/D103766