finagle-core: Fix race in ProcessCoordinate
Summary: Problem

We use a Witness to update the AtomicReference that holds the result of
the `ProcessCoordinate.apply()` function, but that can race with
listeners of the `ProcessCoorindate.changes` Event who are using the
result of `.apply()`, such as Aperture.

Solution

Set the atomic reference directly in the ProcessCoordinate
`setCoordinate` and `unsetCoordinate` functions before notifying the
Event.

JIRA Issues: CSL-5334

Differential Revision: https://phabricator.twitter.biz/D104465