finagle-mux: push-mux client supports opportunistic TLS
Summary: Problem / Solution

Mux is getting support for opportunistic TLS, and the push-based
implementation needs to support it as well.

Notes:
Instead of adding a 'Context' that is propagated down into the
Transporter level, I opted to go with a simple/localized approach of
specializing the PushChannelHandle in the MuxPush object and
downcasting, also in that object. This has the advantage of being simple
and local but is not scalable. That said, we don't need to scale it yet,
and by being local we can easily expand it later. There are two
reasonable alternatives:
- Build the def context: Context into the push-based stack
- Propagate the Stack.Params through the push-based stack

Both of these were rejected since they are general solutions to a local
problem and both can be implemented later, if we need to expand this
idea.

JIRA Issues: CSL-5339

Differential Revision: https://phabricator.twitter.biz/D105634