finagle-mux/finagle-thriftmux: Fail fast on invalid OppTls configuration
Summary: Problem

It's possible to create a mux/thriftmux client or server that has an
illegal opportunistic TLS configuration. Specifically, we can configure
that we want to do opportunistic TLS but don't actually add the security
configuration. This can potentially lead to negotiating TLS but never
upgrading the Netty pipeline, so we then try to interpret ciphertext as
cleartext data which fails at best, and can cause OOM errors at its
worst.

Solution

Fail to create a client of server with an invalid configuration.

JIRA Issues: CSL-5430

Differential Revision: https://phabricator.twitter.biz/D107280