finagle-core: fix usage of epsilon in Ring
Summary: Problem

We defined an epsilon in aperture's ring with the intent to detect
a range of zero when selecting indices in `pick2`. However, we never
normalized its usage which can be dangerous since it can detect
legitimate ranges for sufficiently large ring sizes.

Solution

Remove the usage of epsilon since we don't really need it and can
compare the edge cases directly in both `pick2` and `range`.

JIRA Issues: CSL-5433

Differential Revision: https://phabricator.twitter.biz/D104976