finagle: Remove `NackAdmissionFilter` from protocols without Nack support
Summary: Problem

finagle-memcached, finagle-mysql, and finagle-redis clients do not support
Nacks, so the `NackAdmissionFilter` is unnecessary.

Solution

Remove `NackAdmissionFilter` from the default client stack. Explicitly
add it to the finagle-{http,mux,thriftmux} client stacks and do not add
it to the finagle-{memcached,mysql,redis} client stacks. Note that it is
added to finagle-http2 via finagle-http.

JIRA Issues: CSL-5516

Differential Revision: https://phabricator.twitter.biz/D111722