finagle: Introduce Failure.Ignorable Failure Flag
Summary: Problem

The logic in StatsFilter is coupled to BackupRequestFilter;
it knows about BackupRequestLost in order to omit such
exceptional responses from Stats. We'd like to remove
this coupling.

Solution

Introduce the Failure flag Failure.Ignorable. In StatsFilter,
we now check for this flag when determining whether or
not to stat a response. BackupResponseLost has been
deprecated. Note that BackupRequestFilter still uses
BackupResponseLost, but this filter will soon be replaced
by a new version.

JIRA Issues: CSL-5495

Differential Revision: https://phabricator.twitter.biz/D113466