finagle-memcached: Optimize `PartitioningService.apply` for single-partition requests
Summary: Problem

`PartitioningService.apply` is agnostic to the given request's type. If the
request is a `StorageCommand`, an `ArithmeticCommand`, or a `Delete`, the
request will have just one key and therefore the service will have to speak to
just one partition. `apply` calls `partitionRequest` and `getPartitionKeys` with
these requests even though the returned `Seq`s are discarded. This is wasteful,
especially if a significant fraction of a service's requests are destined for
just one partition.

Solution

Do not call `partitionRequest` if given a `StorageCommand`, an
`ArithmeticCommand`, or a `Delete`.

JIRA Issues: CSL-5517

Differential Revision: https://phabricator.twitter.biz/D114518