finagle-core: create setter for process global load balancer
Summary: Problem

We'd like to be able to set the default load balancer programmatically
so that we can easily configure it based on the environment (e.g. via toggles).

Solution

Create a global var inside the load balancer package which references the
current default and rename the `DefaultBalancerFactory` object to
`FlagBalancerFactory` is more appropriate given its function.

TBR=true

Differential Revision: https://phabricator.twitter.biz/D115335