finagle-mux: Improve representation of pre-encoded messages
Summary: Problem

Our current representation of pre-encoded messages overloads the
Message.buf field to represent the Message.encode(msg) form, which
includes the type and tag, contrary to the definition of the field in
the Message trait.

Solution

Make a new message type, PreEncoded, which caches the encoded
representation of parent messages.

JIRA Issues: CSL-5640

Differential Revision: https://phabricator.twitter.biz/D116547