finagle-netty4: Rename SSL Handlers to Channel Initializers
Summary: Problem

The `Netty4ClientSslHandler` and `Netty4ServerSslHandler` classes are
channel initializers and not descendants of Netty's `SslHandler`. This
can lead to confusion, especially as we move to add our own generic SSL
handler which is a descendant of Netty's `SslHandler`.

Solution

Rename them to `Netty4ClientSslChannelInitializer` and
`Netty4ServerSslChannelInitializer` respectively.

JIRA Issues: CSL-5681

Differential Revision: https://phabricator.twitter.biz/D119375