finagle: Only add `NackAdmissionFilter` to protocols with Nack support
Summary: Problem

finagle-memcached, finagle-mysql, and finagle-redis clients do not support
Nacks, so the `NackAdmissionFilter` is unnecessary.

Solution

Remove `NackAdmissionFilter` from the default client stack. Explicitly add it to
the finagle-{http,mux} client stacks and do not add it to the
finagle-{memcached,mysql,redis} client stacks. Note that it is added to
finagle-http2 via finagle-http and finagle-thriftmux via finagle-mux.

JIRA Issues: csl-5641

Differential Revision: https://phabricator.twitter.biz/D116722