finagle-{core, memcached}: remove `ConcurrentLoadBalancerFactory` in favor of stack param.
Summary: Problem

It can be useful to duplicate addresses inside the load balancer as
to increase concurrency. For example, this can remove head-of-line
blocking for pipelined protocols or when servers pin connections to
a thread for their lifetime.

Solution

`finagle-memcached` had this built in via a `LoadBalancerFactory` extension.
We remove that extension in favor of a simple stack param inside the LBF.

JIRA Issues: CSL-5679

Differential Revision: https://phabricator.twitter.biz/D119394