finagle-http2: Downgrade correctly with ALPN
Summary: Problem

Right now our ALPN downgrade drops the old connection on the ground and then
makes a new http/1.1 connection

Solution

Instead of dropping the old connection on the ground, change the pipeline to
http/1.1 and continue to use it.

Result

We no longer leak connections when the ALPN upgrade fails.

JIRA Issues: CSL-5425

Differential Revision: https://phabricator.twitter.biz/D118687