finagle-http2: Bump the default header list size
Summary: Problem

Folks see increased rates of 431s when migrating from http/1.1 to http/2.
This seems to be because of a bug in netty which miscounts the size of headers
by double counting header names.

See https://github.com/netty/netty/issues/7511.

Solution

The longterm fix is to fix netty, but the short term is to bump the max header
list size.

JIRA Issues: CSL-5725

Differential Revision: https://phabricator.twitter.biz/D121912