finagle-memcached: Speedup Buf to Int conversion
Summary: Problem

The memcached protocol often returns lengths encoded as bytes, encoded
as ascii, "101" for example. These need to be parsed into their
integer form. The bytes are in Bufs, and the code is writing them to a
byte array to then be parsed.

Solution

Skip the middle man, the byte array, and parse the Buf directly.

Result

A 7.4% runtime reduction in a benchmark for decoding GETs.

JIRA Issues: CSL-5775

Differential Revision: https://phabricator.twitter.biz/D125543