finagle-base-http: Introduce CookieCodec abstraction
Summary: Problem

We'd like to add a Toggle to switch from using the Netty 3 cookie
encoder/decoder to Netty 4. In order to transparently swap
out encoders/decoders, we need them to have a common
interface.

Solution

Introduce CookieCodec, which provides methods to
encode/decode, and an implementation that uses the
Netty 3 versions of encoders/decoders.

JIRA Issues: CSL-5771

Differential Revision: https://phabricator.twitter.biz/D124988