finagle-core: Increase the min Deterministic Aperture size to 8
Summary: Problem

We've observed that we can get 4 nodes down in a Deterministic Aperture,
and since we don't have a mechanism for evicting unhealthy nodes for
this distributor, we effectively kill the instance that depends on this
client.

Solution

To decrease the probability of finding a completely dead set of nodes we
increase the minimum to 8. This will be revisited as we consider the
case of allowing the Deterministic Aperture to adjust for unhealthy
nodes in its slice.

JIRA Issues: CSL-5746

Differential Revision: https://phabricator.twitter.biz/D123187