finagle-mysql: Add failure that caused rollback to log of rollback failure
Summary: Problem

When a transaction fails and is rolled back, we log if
the rollback fails and include the failure that caused
the rollback to fail. However, the original failure is not
logged and it's confusing to users whether the logged
failure is the one that caused the rollback, or the one
that caused the rollback to fail.

Solution

Include the failure that caused the rollback in the log.

JIRA Issues: CSL-5794

Differential Revision: https://phabricator.twitter.biz/D126350