finagle/finagle-mysql: Don't roll back a transaction if the failure was a wrapped ChannelClosedException
Summary: Problem

Transactions are rolled back even if the failure was a
ChannelClosedException. This is unnecessary because
if the channel is closed, the rollback will also fail with
a CCE.

Solution

Don't roll back a transaction if the failure was a wrapped
ChannelClosedException. We need to unwrap the exception
because currently, CCEs are wrapped in a WriteException
(and there could potentially be more levels of wrapping). We
also no longer log transaction rollback failures if the failure
was a wrapped CCE, because this cannot be mitigated by
the user.

Differential Revision: https://phabricator.twitter.biz/D126526