finagle-core: Rename Aperture.aperture to Aperture.logicalAperture
Summary: Problem / Solution

We have two notions of what an aperture is, a logical and a physical
aperture (in the case of vanilla Aperture they are the same) but we have
a lot of things named 'aperture' which can be confusing, especially the
stats that we emit.

Solution

Rename the things called 'aperture' to 'logicalAperture' (or 'logical_aperture'
for the stat) so as to be clear which one we're working with.

JIRA Issues: CSL-5752

Differential Revision: https://phabricator.twitter.biz/D126357