finagle-http2: Close http2 connections on failure detection
Summary: Problem:
HTTP/2 supports Ping messages, which are a great way to detect remote health
and a natural candidate for use with `FailureDetector`. However, this is
currently disabled on HTTP/2 clients because our connection pool doesn't
understand how H2 streams work. The result is that if `FailureDetector`
marks a connection as `Closed`, no requests are sent along that connection,
but the connection is never torn down.

Solution:
Re-enabled `FailureDetector` and tear down the connection when it triggers.

JIRA Issues: CSL-4498

Differential Revision: https://phabricator.twitter.biz/D126840