Don't calculate the hashcode of Stream in Budget
Calculating the hashcode of an instance of Retries.Budget could end up
in an infinite computation.
Since the case class gave us a default implementation including all
fields, including the requeueBackoffs field with a possibly infinite
Stream[Duration], we need to override this behaviour.

This is done by ignoring the fact that there is a field with a Stream.
The equality of the class is implemented with
http://www.artima.com/lejava/articles/equality.html in mind.

Closes #669.