finagle-toggle: Make Toggles independent
Summary: Problem

Toggles all use the same hash seed (a constant integer), so
that when the server id is applied to a toggle to produce an
on/off value, toggles with the same fraction have the same
distribution and thus have the same on/off state.

Solution

Change the constant integer hash seed to be a hash of the
Toggle id.

JIRA Issues: CSL-5817

Differential Revision: https://phabricator.twitter.biz/D128172