finagle-core: Don't Calculate the Hashcode of Stream in Budget
Summary: Problem / Solution

Calculating the hashcode of an instance of `Retries.Budget` could end up
in an infinite computation. Since the case class gives us a default
implementation including all fields, including the `requeueBackoffs` field
with a possibly infinite `Stream[Duration]`, we need to override this behaviour.
This is done by ignoring the fact that there is a field with a `Stream`.
The equality of the class is implemented with
http://www.artima.com/lejava/articles/equality.html in mind.

Closes #669.

Signed-off-by: Ryan O'Neill <ryano@twitter.com>

JIRA Issues: CSL-5823

Differential Revision: https://phabricator.twitter.biz/D128594