finagle-netty4-http: Netty4CookieCodec that uses Netty 4 Encoder/Decoder
Summary: Problem

We want to switch from Netty 3 to Netty 4 encoding/decoding
for Cookies. Thus, we need an implementation of CookieCodec
that uses a Netty 4 encoder/decoder.

Solution

Introduce Netty4CookieCodec. Note that this codec is not yet
used; we will switch over from Netty3CookieCodec using a
toggle (will be wired in in a future commit).

JIRA Issues: CSL-5790

Differential Revision: https://phabricator.twitter.biz/D127954