finagle-http2: Don't get in an RST loop
Summary: Problem

We sometimes send an RST after we receive an RST, which is explicitly disallowed
by the spec.  Although we don't seem to be getting into RST loops, it's possible
if both the client and the server are badly behaved, so we should avoid having
badly behaved clients.

Solution

Mark the transport as Dead so that it knows it doesn't need to RST.

JIRA Issues: CSL-5879

Differential Revision: https://phabricator.twitter.biz/D130963