finagle-http2: Fix memory leak in AdapterProxyChannelHandler
Summary: Problem

The AdapterProxyChannelHandler is potentially leaking reference counted
memory when a race occurs between the client resetting a stream and the
peer sending more stream data, in which case the AdapterProxyChannelHandler
just discards the message where it should be releasing it.

Solution

Release the message under these circumstances.

JIRA Issues: CSL-5899

Differential Revision: https://phabricator.twitter.biz/D132315