finagle-mux: Fix potential NPE in c.t.f.m.e.pushsession.ServerTracker
Summary: Problem

In the `ServerTracker.renderResponse` method, we check that the wrong
pointer is not null in ServerTracker which could result in an NPE since
the pointer that could  be null could end up being dereferenced.

Solution

Restructure the method into a match block which allows us to avoid
naming the pointer until after we've done the null check.

Differential Revision: https://phabricator.twitter.biz/D132698