finagle: Move toggles from finagle-http to finagle-base-http
Summary: Problem

Users may test Requests and Responses independently of
using any finagle-http classes (Requests and Responses are in
finagle-base-http). This means that toggles in finagle-http are not
evaluated in the test run; specifically, the toggle for the cookie encoding
codec.

Solution

Move toggles in finagle-http to finagle-base-http.

JIRA Issues: CSL-5900

Differential Revision: https://phabricator.twitter.biz/D132162