finagle-core: Don't use Updater in Balancer
Summary: Problem

We use the Updater primitive to manage concurrency in our Balancer
implementation but are not using it right. For example, we may call
`Balancer.rebuild()` and then ask for a new node from the new
distributor but we may not actually be the one to run the `rebuild` so
we actually request a node from the old `Distributor`. Measurements show
that operations on `Balancer` are almost never run concurrently so it's not
providing any meaningful benefit through its event stream pre-processing
logic.

Solution

Remove the `Updater` and use plain old synchronization for rebuild and
update operations. This reduces complexity and makes our usages of
`rebuild` correct.

JIRA Issues: CSL-5751

Differential Revision: https://phabricator.twitter.biz/D126486