finagle-core: Bump the minimum aperture for D-Aperture to 12
Summary: Problem

It has been observed that finagle clients using the
determinstic-aperture load balancer have an increase in latency,
especially during the deployment of backends, in which case a number of
nodes are expected to be down.

Solution

Increase the minimum aperture from 8 to 12 to allow for more healthy
connections in the aperture during backend deployments.

JIRA Issues: CSL-5915

Differential Revision: https://phabricator.twitter.biz/D132434