finagle-base-http: Cookie.maxAge is DefaultMaxAge if set to None
Summary: Problem

In Cookie's copy constructor, we can set maxAge to None,
but encoding to Netty will throw an NPE when we try to
call Cookie.maxAge.inSeconds on a null max-age. The other getters
are fine returning null because no methods are called on the results.

Solution

Return Cookie.DefaultMaxAge if maxAge is None.

JIRA Issues: CSL-6028

TBR=true

Differential Revision: https://phabricator.twitter.biz/D138393