finagle-netty4: Use SafeRunnable to cut some allocs in Netty4PushChannelHandle
Summary: Problem

In the c.t.f.n.e.p.Netty4PushChannelHandle we proxy all `Runnable`
instances to ensure that if they throw an exception we log loudly and the
channel is closed to prevent resource leaks. This results in an extra
`Runnable` allocation on many operations.

Solution

Make a new type, SafeRunnable, that only lives in the
Netty4PushChannelHandle and can be used directly for operations within
the handle, avoiding the proxy instance.

JIRA Issues: CSL-6050

Differential Revision: https://phabricator.twitter.biz/D139732