finagle-netty4-http: Fix bug in Netty4FormPostEncoder
Summary: Problem

The Netty4FormPostEncoder can generate a request with a content that is
backed by an underlying Netty4 ByteBuf, which gets unwrapped and
decremented in the Netty encoding logic. This can lead to reference
count exceptions if the request is reused.

Solution

Make sure we don't emit any ByteBufAsBuf instances.

JIRA Issues: CSL-6068

Differential Revision: https://phabricator.twitter.biz/D140853