finagle-mux/finagle-http2: ClientDiscardedRequestException shouldn't have a stack trace
Summary: Problem

There is little value in the stack trace for
ClientDiscardedRequestException since it doesn't represent an actual
failure: it's purpose is to provide a notice that work being performed
by a server is no longer necessary. Since this is often logged, we end
up diminishing the value of logs by saturating them with these stack
traces.

Solution

Make the ClientDiscardedRequestException extend the NoStackTrace trait
to suppress stack trace generation.

JIRA Issues: CSL-6069

Differential Revision: https://phabricator.twitter.biz/D141184