finagle-zipkin-scribe: Renamed finagle-zipkin to finagle-zipkin-scribe
Summary: Problem

finagle-zipkin was originally named when scribe was the only transport
for zipkin, so zipkin structs over scribe was the zipkin protocol. That's no
longer the case--there are many different possible transports, and the
openzipkin folks want to remove the scribe transport.

Solution

Rename finagle-zipkin to finagle-zipkin-scribe so it's clear it's just an
implemenation, instead of the default.  This will discourage folks from adopting
it by default.

JIRA Issues: CSL-6090

Differential Revision: https://phabricator.twitter.biz/D141940