finagle-thrift: Move stateless `ThriftRichClient` methods to a new trait
Summary: Problem

Certain methods in the `c.t.finagle.thrift.ThriftRichClient` are stateless
in that they work only over the inputs, however because they are defined
on the Thrift or ThriftMux client class, they look like they work over the
mutated params of the class, when they do not.

Solution

Move the stateless methods which are mere functions to wrap or convert the
input parameters to a companion trait which is mixed into the Client class
companion.

Result

Less confusion about these methods which are mere convienence methods and
not stack methods.

JIRA Issues: CSL-6108

TBR=true

Differential Revision: https://phabricator.twitter.biz/D143185