finagle-memcached: MockClient now understands expiry and flags
Summary: Problem

As a user of MockClient I want to be able to make assertions about the
expiry time that is associated with items written to cache. The
current MockClient ignores these values, forcing users to use Mockito
to make assertions.

Solution

`MockClient` now stores wrapped values in a mutable `Map`. The wrapper
contains the expiry `Time`, flags `Int`, and value `Buf`. On access,
the expiry time is checked, and if the item has expired, the mock acts
as if the item is missing.

JIRA Issues: CACHE-12277

TBR=true

Differential Revision: https://phabricator.twitter.biz/D148862