finagle-redis: Redis cluster management
Summary: Problem

The current finagle-redis client does not have support for the CLUSTER
commands introduced in Redis 3.0.0.

Solution

This adds support for a number of new cluster commands that are
necessary to setup and manage a redis cluster.

- CLUSTER ADDSLOT, assigns a storage slot to a server
- CLUSTER SETSLOT, updates the state of a slot
- CLUSTER SLOTS, returns a list of slots known to the server
- CLUSTER GETKEYSINSLOT, returns keys stored in a slot
- CLUSTER REPLICATE, changes a node to become a replica of a server
- CLUSTER MEET, introduces cluster nodes to each other
- CLUSTER NODES, returns the list of cluster nodes known to the server
- CLUSTER INFO, returns the cluster specific info
- MIGRATE, used to transfer keys to another node

Result

Using the introduced commands it is possible to setup a redis cluster
for testing and live re-sharding. This is the first step in
implementing the redis cluster client protocol. The next step is to
handle operation redirects, i.e. when the server returns ASK or MOVED.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D152186