finagle-netty4: Track ByteBuf leaks by default
Summary: Problem

We really want to know if we're leaking a buffer, but we don't record
that via metrics by default, limiting our visibility to the logs Netty
emits, which can be tough to monitor on a large scale.

Solution

Restructure the counter so that it isn't created unless we actually
detect a leak and turn on leak detection by default.

JIRA Issues: CSL-5964

Differential Revision: https://phabricator.twitter.biz/D152828