finagle-serversets: Preserve host/ip provided in serveIface
Summary: Problem

I'm deploying Finagle ThriftMux with Zookeeper ServerSets inside
Kubernetes and pod hostname is being set as the host in Zookeeper
instead of the IP of the pod which is what I'm binding to when initing
the ThriftMux server.

Solution

In k8s getHostName returns the name of the pod which is not accessible
from other pods in the cluster. getHostString returns the IP of the
pod.

Fixes #684

Signed-off-by: Kevin Oliver <koliver@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D153429