finagle-http2: Oversized headers result in TooLongMessageException
Summary: Problem

If the remote server returns a header list that's too big, we just swallow the
error.  It would be better if we matched h1 behavior, which is to hand back
a TooLongMessageException.

Solution

Whenever Http2ConnectionHandler#onStreamError is called with an
HeaderListSizeException, match http/1.1 behavior by handing back a
TooLongMessageException

JIRA Issues: CSL-5810

Differential Revision: https://phabricator.twitter.biz/D154113