finagle-base-http: Provide Uri for doing URI things
Summary: Problem

The c.t.f.http.Request class has a hodge-podge of things on it, and it's often
difficult to divine what the different pieces are doing.  As an example, the
`Request#params` refer to things that don't come from the HTTP spec, URI query
parameters and the application/x-www-form-urlencoded MIME-type.

Solution

Start breaking out the non-HTTP pieces of `c.t.f.http.Message` and friends so
that in the long run we can slim down `c.t.f.http.Message` to something
reasonable.

Result

It's now easy to extract the URI parameters as a ParamMap.

JIRA Issues: CSL-6251

Differential Revision: https://phabricator.twitter.biz/D160298