finagle-core, finagle-http: Add ability to suppress ServerAC
Summary: Problem

ServerAdmissionControl is responsible for nacking requests but it does
so without any context as to whether the client could even retry it if
it wanted to. For example, a HTTP client may be streaming a body from a
Reader that it didn't generate directly, so if that request gets
nacked, there is no way to get back any data that was read during the
first dispatch, and therefore the request couldn't be retried even if
it is known the server didn't process it.

Solution

Create a local context value that acts as a signal for
ServerAdmissionControl to interpret as 'this cannot be retried, so don't
nack this request'. Since HTTP is our only protocol that suffers from this
property (thus far) HTTP servers will signal that all requests that have
a body shouldn't be nacked if they are either
1. chunked- our HTTP/1.x server requires chunked requests be consumed
    which we can't do in the nack filter, resulting in a borked session.
2. have a body and don't have the magic header sent by the client that
    signals that it is able to retry the request body.

JIRA Issues: CSL-4614

Differential Revision: https://phabricator.twitter.biz/D134209